+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 
 - Use 5 minutes recording chunk size as default
 - Protect against running out of disk space
+- Add uploader based on ``rsync``
 
 
 2021-06-20 0.2.0